Drop passport.
Passport was too complex and lacking documentation, resulting in many
hours to reverse engineer many layers of frameworks and pluging, to find
out what it was doing or could do. And ultimately it wasn't capable of
per-request callback urls, which are required to support many websites
from a single server.

After all, OAuth is simple. Only two requests to handle: to initiate
authentication, a redirect to the authentication endpoint, with a few
parameters; and after the callback POST request to exchange access code
for access and id tokens. It doesn't require a dozen inscrutable,
undocumented packages.

This rewrite is a bit crude, but it is working and callback URL is
configured in nginx as part of the site configuration. Each site can
have its own callback. It is also trivially easy to see what is
happening, as opposed to the opaque passport calls.